Don't run training is model is already present

+++ pipeline/train/finetune-student.sh
@@ -21,20 +21,21 @@ test -v SRC
 test -v TRG
 test -v WORKDIR
 
-mkdir -p "${dir}"
-cp "${student}/model.npz.best-bleu-detok.npz" "${dir}/model.npz"
-cp "${student}/vocab.spm" "${dir}/"
+if [ ! -s "${dir}/model.npz.best-bleu-detok.npz" ]; then
+  mkdir -p "${dir}"
+  cp "${student}/model.npz.best-bleu-detok.npz" "${dir}/model.npz"
+  cp "${student}/vocab.spm" "${dir}/"
 
-bash "${WORKDIR}/pipeline/train/train.sh" \
-  "${WORKDIR}/pipeline/train/configs/model/student.tiny11.yml" \
-  "${WORKDIR}/pipeline/train/configs/training/student.finetune.yml" \
-  "${SRC}" \
-  "${TRG}" \
-  "${corpus}" \
-  "${devset}" \
-  "${dir}" \
-  --guided-alignment "${alignment}/corpus.aln.gz"
-
+  bash "${WORKDIR}/pipeline/train/train.sh" \
+    "${WORKDIR}/pipeline/train/configs/model/student.tiny11.yml" \
+    "${WORKDIR}/pipeline/train/configs/training/student.finetune.yml" \
+    "${SRC}" \
+    "${TRG}" \
+    "${corpus}" \
+    "${devset}" \
+    "${dir}" \
+    --guided-alignment "${alignment}/corpus.aln.gz"
+fi
 
 echo "###### Done: Finetuning the student model"
